Sprint 6/Desafio/desafio-1/requirements.txt
boto3
python-dotenv
datetime